requirements.txt
# pip install mcp aiohttp python-dotenv orjson
# Core MCP package
mcp>=1.0.0

# For HTTP requests to your API
aiohttp>=3.8.0

# For environment variable management
python-dotenv>=1.0.0

# Optional: For better JSON handling
orjson>=3.8.0